// src/generated-api/src/models/HandlersPerformanceSummaryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DomainPerformanceSummary } from './DomainPerformanceSummary';
import {
    DomainPerformanceSummaryFromJSON,
    DomainPerformanceSummaryFromJSONTyped,
    DomainPerformanceSummaryToJSON,
    DomainPerformanceSummaryToJSONTyped,
} from './DomainPerformanceSummary';
import type { HandlersDateRangeInfo } from './HandlersDateRangeInfo';
import {
    HandlersDateRangeInfoFromJSON,
    HandlersDateRangeInfoFromJSONTyped,
    HandlersDateRangeInfoToJSON,
    HandlersDateRangeInfoToJSONTyped,
} from './HandlersDateRangeInfo';

/**
 * 
 * @export
 * @interface HandlersPerformanceSummaryResponse
 */
export interface HandlersPerformanceSummaryResponse {
    /**
     * 
     * @type {DomainPerformanceSummary}
     * @memberof HandlersPerformanceSummaryResponse
     */
    data?: DomainPerformanceSummary;
    /**
     * 
     * @type {HandlersDateRangeInfo}
     * @memberof HandlersPerformanceSummaryResponse
     */
    dateRange?: HandlersDateRangeInfo;
    /**
     * 
     * @type {string}
     * @memberof HandlersPerformanceSummaryResponse
     */
    status?: string;
}

/**
 * Check if a given object implements the HandlersPerformanceSummaryResponse interface.
 */
export function instanceOfHandlersPerformanceSummaryResponse(value: object): value is HandlersPerformanceSummaryResponse {
    return true;
}

export function HandlersPerformanceSummaryResponseFromJSON(json: any): HandlersPerformanceSummaryResponse {
    return HandlersPerformanceSummaryResponseFromJSONTyped(json, false);
}

export function HandlersPerformanceSummaryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): HandlersPerformanceSummaryResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : DomainPerformanceSummaryFromJSON(json['data']),
        'dateRange': json['dateRange'] == null ? undefined : HandlersDateRangeInfoFromJSON(json['dateRange']),
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function HandlersPerformanceSummaryResponseToJSON(json: any): HandlersPerformanceSummaryResponse {
    return HandlersPerformanceSummaryResponseToJSONTyped(json, false);
}

export function HandlersPerformanceSummaryResponseToJSONTyped(value?: HandlersPerformanceSummaryResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': DomainPerformanceSummaryToJSON(value['data']),
        'dateRange': HandlersDateRangeInfoToJSON(value['dateRange']),
        'status': value['status'],
    };
}