// src/generated-api/src/models/HandlersPerformanceTimeSeriesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DomainPerformanceTimeSeriesPoint } from './DomainPerformanceTimeSeriesPoint';
import {
    DomainPerformanceTimeSeriesPointFromJSON,
    DomainPerformanceTimeSeriesPointFromJSONTyped,
    DomainPerformanceTimeSeriesPointToJSON,
    DomainPerformanceTimeSeriesPointToJSONTyped,
} from './DomainPerformanceTimeSeriesPoint';

/**
 * 
 * @export
 * @interface HandlersPerformanceTimeSeriesResponse
 */
export interface HandlersPerformanceTimeSeriesResponse {
    /**
     * 
     * @type {Array<DomainPerformanceTimeSeriesPoint>}
     * @memberof HandlersPerformanceTimeSeriesResponse
     */
    data?: Array<DomainPerformanceTimeSeriesPoint>;
    /**
     * 
     * @type {string}
     * @memberof HandlersPerformanceTimeSeriesResponse
     */
    status?: string;
}

/**
 * Check if a given object implements the HandlersPerformanceTimeSeriesResponse interface.
 */
export function instanceOfHandlersPerformanceTimeSeriesResponse(value: object): value is HandlersPerformanceTimeSeriesResponse {
    return true;
}

export function HandlersPerformanceTimeSeriesResponseFromJSON(json: any): HandlersPerformanceTimeSeriesResponse {
    return HandlersPerformanceTimeSeriesResponseFromJSONTyped(json, false);
}

export function HandlersPerformanceTimeSeriesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): HandlersPerformanceTimeSeriesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(DomainPerformanceTimeSeriesPointFromJSON)),
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function HandlersPerformanceTimeSeriesResponseToJSON(json: any): HandlersPerformanceTimeSeriesResponse {
    return HandlersPerformanceTimeSeriesResponseToJSONTyped(json, false);
}

export function HandlersPerformanceTimeSeriesResponseToJSONTyped(value?: HandlersPerformanceTimeSeriesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(DomainPerformanceTimeSeriesPointToJSON)),
        'status': value['status'],
    };
}