// src/generated-api/src/models/HandlersDailyPerformanceReportResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DomainDailyPerformanceReport } from './DomainDailyPerformanceReport';
import {
    DomainDailyPerformanceReportFromJSON,
    DomainDailyPerformanceReportFromJSONTyped,
    DomainDailyPerformanceReportToJSON,
    DomainDailyPerformanceReportToJSONTyped,
} from './DomainDailyPerformanceReport';
import type { DomainPaginationResult } from './DomainPaginationResult';
import {
    DomainPaginationResultFromJSON,
    DomainPaginationResultFromJSONTyped,
    DomainPaginationResultToJSON,
    DomainPaginationResultToJSONTyped,
} from './DomainPaginationResult';

/**
 * 
 * @export
 * @interface HandlersDailyPerformanceReportResponse
 */
export interface HandlersDailyPerformanceReportResponse {
    /**
     * 
     * @type {Array<DomainDailyPerformanceReport>}
     * @memberof HandlersDailyPerformanceReportResponse
     */
    data?: Array<DomainDailyPerformanceReport>;
    /**
     * 
     * @type {DomainPaginationResult}
     * @memberof HandlersDailyPerformanceReportResponse
     */
    pagination?: DomainPaginationResult;
    /**
     * 
     * @type {string}
     * @memberof HandlersDailyPerformanceReportResponse
     */
    status?: string;
}

/**
 * Check if a given object implements the HandlersDailyPerformanceReportResponse interface.
 */
export function instanceOfHandlersDailyPerformanceReportResponse(value: object): value is HandlersDailyPerformanceReportResponse {
    return true;
}

export function HandlersDailyPerformanceReportResponseFromJSON(json: any): HandlersDailyPerformanceReportResponse {
    return HandlersDailyPerformanceReportResponseFromJSONTyped(json, false);
}

export function HandlersDailyPerformanceReportResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): HandlersDailyPerformanceReportResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(DomainDailyPerformanceReportFromJSON)),
        'pagination': json['pagination'] == null ? undefined : DomainPaginationResultFromJSON(json['pagination']),
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function HandlersDailyPerformanceReportResponseToJSON(json: any): HandlersDailyPerformanceReportResponse {
    return HandlersDailyPerformanceReportResponseToJSONTyped(json, false);
}

export function HandlersDailyPerformanceReportResponseToJSONTyped(value?: HandlersDailyPerformanceReportResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(DomainDailyPerformanceReportToJSON)),
        'pagination': DomainPaginationResultToJSON(value['pagination']),
        'status': value['status'],
    };
}